backend/requirements.txt
fastapi
uvicorn[standard]
pillow
transformers
torch
boto3
botocore
python-multipart
